requirements.txt
tensorflow>=2.12
numpy
pandas
matplotlib
seaborn
scikit-learn
streamlit
Pillow
